+++ src/reference-data/utils.ts
@@ -7,6 +7,14 @@ import {
 import {APP_VERSION} from '@env';
 import {compareVersion} from '@atb/utils/compare-version';
 import {CustomerProfile} from '@atb/ticketing';
+import {
+  FareContractTexts,
+  TranslateFunction,
+  useTranslation,
+} from '@atb/translations';
+import {FareProductTypeConfig} from '@atb/configuration';
+import {TransportModePair} from '@atb/components/transportation-modes';
+import {useTextForLanguage} from '@atb/translations/utils';
 
 /**
  * Wrapper for getting the name of a NeTeX entity in the given language.
@@ -58,4 +66,50 @@ export const isProductSellableInApp = (
     return true;
 
   return product.distributionChannel.some((channel) => channel === 'app');
+};
+
+const modesDisplayLimit = 2;
+
+export const useFareProductTitle = (config?: FareProductTypeConfig) => {
+  const {t} = useTranslation();
+
+  const transportModes = config?.transportModes;
+
+  const transportModesText = getFareProductTravelModesText(
+    transportModes,
+    t,
+    modesDisplayLimit,
+  );
+
+  return (
+    useTextForLanguage(config?.name) +
+    (transportModesText ? ', ' + transportModesText : '')
+  );
+};
+
+const getFareProductTravelModesText = (
+  modes: TransportModePair[] | undefined,
+  t: TranslateFunction,
+  modesDisplayLimit = 2,
+): string | undefined => {
+  if (!modes) {
+    return undefined;
+  }
+  if (modes.length > modesDisplayLimit) {
+    return t(FareContractTexts.transportModes.multipleTravelModes);
+  }
+
+  const travelModes = modes
+    .map((tm) => t(FareContractTexts.transportMode(tm.mode, tm.subMode)))
+    .filter((value, index, array) => array.indexOf(value) === index); // remove duplicates
+
+  if (travelModes.length < 2) {
+    return travelModes?.[0];
+  } else {
+    return (
+      travelModes.splice(0, travelModes.length - 1).join(', ') +
+      ` ${t(FareContractTexts.transportModes.concatListWord)} ` + // add " and " between the last two
+      travelModes
+    );
+  }
 };
