+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/index.ts
@@ -1,2 +1,1 @@
 export {TicketTabNav_PurchaseTabScreen} from './TicketTabNav_PurchaseTabScreen';
-export {useFareProductTitle} from './use-fare-product-title';
